+++ threedigrid_builder/application.py
@@ -68,6 +68,9 @@ def _make_gridadmin(
         # Patch epsg code with that of the DEM (so: user-supplied EPSG is ignored)
         grid.meta.epsg_code = raster.epsg_code
 
+        # Use raster epsg_code instead of the one from global_settings
+        db.epsg_code = raster.epsg_code
+
         refinements = db.get_grid_refinements()
         progress_callback(0.7, "Constructing 2D computational grid...")
         quadtree = QuadTree(
@@ -82,7 +85,15 @@ def _make_gridadmin(
             node_id_counter=node_id_counter,
             line_id_counter=line_id_counter,
         )
+
+        if grid.meta.has_groundwater:
+            grid.add_groundwater_nodes(grid.nodes, node_id_counter)
+            grid.add_groundwater_vertical_lines(grid.nodes, line_id_counter)
+
         grid.set_obstacles(db.get_obstacles(), db.get_levees())
+        if grid.meta.has_groundwater_flow:
+            grid.add_groundwater_lines(grid.lines, line_id_counter)
+
         grid.set_boundary_conditions_2d(
             db.get_boundary_conditions_2d(),
             quadtree,