+++ CHANGES.rst
@@ -1,8 +1,22 @@
 Changelog of threedigrid-builder
 ================================
 
-0.8.4 (unreleased)
+0.9.1 (unreleased)
 ------------------
+
+- Bugfix: use DEM epsg_code for 2D models.
+
+- Bugfix: Small fix for lgrtot.
+
+- Bugfix: Small fix adding groundwater cells.
+
+
+0.9.0 (2021-12-15)
+------------------
+
+- Add padding to area_mask for creating quadtree.
+
+- Added groundwater 2D nodes, 2D vertical lines, and 2D groundwater lines.
 
 - Bugfix: cross section tabulate used wrong width/height.
 
@@ -19,6 +33,7 @@ Changelog of threedigrid-builder
   result in an incorrect cast to a numpy object dtype.
 
 - Fixed bug with zero-d administration `cci` (index needs to be 1-based), removed `cid` field.
+
 
 0.8.2 (2021-12-05)
 ------------------